+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - output: add FileWriter output in replacement of Files output  [#31](https://github.com/AdRoll/baker/pull/31)
 - upload: s3: add `ExitOnError` configuration [#27](https://github.com/AdRoll/baker/pull/27)
 - uploads now return an error instead of panicking and baker deals with it [#27](https://github.com/AdRoll/baker/pull/27)
+- general: replace `${KEY}` in the TOML conf with the `$KEY` env var [#24](https://github.com/AdRoll/baker/pull/24)
 
 ### Changed